* added support for sourcing files from HTTP links in the wxs.template file: the gene2protein.db3 file is now included when the IDPicker 3 installer is built * fixed cross-thread call to Close() * added memory-mapped IO to Qonverter and tweaked query performance - enabled memory-mapped IO in file import step - fixed unresponsive tree grouping bug in the protein view - added "Modified Sites" column to protein view, which shows a list of mods on that protein (or first protein of the protein group) like "80@T422" - default tree grouping mode for protein view is now Protein Group - fixed -ApplySourceGroupHierarchy command-line option - added a message box prompting user to embed gene metadata to new databases: I will make this user-configurable later and right now it doesn't actually disable the gene features, but it will do the embedding - added ReporterIonMzTolerance parameter to idpQonvert; will add later to the GUI
Former-commit-id: 6483d080f4a52775d7b79ccb2b92a46c1c974836

+++ pwiz_tools/Bumbershoot/idpicker/Program.cs
@@ -131,7 +131,7 @@ namespace IDPicker
         {
             if (MainWindow == null)
             {
-                if (IsHeadless)
+                if (IsHeadless && !System.Diagnostics.Debugger.IsAttached)
                 {
                     Console.Error.WriteLine("\r\nError: {0}\r\n\r\nDetails:\r\n{1}", e.Message, e.ToString());
                     System.Diagnostics.Process.GetCurrentProcess().Kill();
@@ -157,7 +157,7 @@ namespace IDPicker
             {
                 reportForm.StartPosition = FormStartPosition.CenterParent;
 
-                if (IsHeadless)
+                if (IsHeadless && !System.Diagnostics.Debugger.IsAttached)
                 {
                     Console.Error.WriteLine("\r\nError: {0}\r\n\r\nDetails:\r\n{1}", reportForm.ExceptionType, reportForm.MessageBody);
                     System.Diagnostics.Process.GetCurrentProcess().Kill();
